requirements.txt
line-bot-sdk
flask
requests
bs4
lxml
googletrans
pafy
youtube-dl
pytube